# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Our Empire - Satta Matka Application API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/dashboard:
    get:
      operationId: getUserDashboard
      tags: [user]
      summary: Get user dashboard data
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDashboard"
  /user/balance:
    get:
      operationId: getUserBalance
      tags: [user]
      summary: Get user balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceResponse"
  /user/bets:
    get:
      operationId: getUserBets
      tags: [user]
      summary: Get user bets
      security:
        - bearerAuth: []
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [today, yesterday, week, all]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bets list
          content:
            application/json:
              schema:
                type: object
                properties:
                  bets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bet"
                required: [bets]
  /user/winnings:
    get:
      operationId: getUserWinnings
      tags: [user]
      summary: Get user winnings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Winnings
          content:
            application/json:
              schema:
                type: object
                properties:
                  winnings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Bet"
                  totalWon:
                    type: number
                required: [winnings, totalWon]
  /user/transactions:
    get:
      operationId: getUserTransactions
      tags: [user]
      summary: Get user transactions
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
                required: [transactions]
  /user/referrals:
    get:
      operationId: getUserReferrals
      tags: [user]
      summary: Get referral info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReferralInfo"
  /user/analytics:
    get:
      operationId: getUserAnalytics
      tags: [user]
      summary: Get user analytics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAnalytics"
  /user/change-password:
    post:
      operationId: changePassword
      tags: [user]
      summary: Change password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                oldPassword:
                  type: string
                newPassword:
                  type: string
              required: [oldPassword, newPassword]
      responses:
        "200":
          description: Changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /user/notifications:
    get:
      operationId: getUserNotifications
      tags: [user]
      summary: Get user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: "#/components/schemas/Notification"
                  unreadCount:
                    type: integer
                required: [notifications, unreadCount]

  # MARKETS
...
  /wallet/add-money-request:
    post:
      operationId: addMoneyRequest
      tags: [wallet]
      summary: Submit add money request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMoneyRequest"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/withdrawal-request:
    post:
      operationId: withdrawalRequest
      tags: [wallet]
      summary: Submit withdrawal request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequest"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/history:
    get:
      operationId: getWalletHistory
      tags: [wallet]
      summary: Get wallet history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactions:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
                required: [transactions]
  /wallet/upi-details:
    get:
      operationId: getUpiDetails
      tags: [wallet]
      summary: Get assigned UPI details for user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: UPI details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpiDetails"

  # RESULTS
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required: [message, success]
...
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
      required: [balance]
    UserDashboard:
      type: object
      properties:
        balance:
          type: number
        totalBets:
          type: integer
        totalWon:
          type: number
        totalLost:
          type: number
        recentBets:
          type: array
          items:
            $ref: "#/components/schemas/Bet"
        activeMarkets:
          type: array
          items:
            $ref: "#/components/schemas/Market"
        recentResults:
          type: array
          items:
            $ref: "#/components/schemas/Result"
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadNotifications:
          type: integer
      required: [balance, totalBets, totalWon, totalLost, recentBets, activeMarkets, recentResults]
    UserAnalytics:
      type: object
      properties:
        totalBets:
          type: integer
        totalWagered:
          type: number
        totalWon:
          type: number
        winRate:
          type: number
        profitLoss:
          type: number
        favoriteMarket:
          type: string
        monthlyData:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              wagered:
                type: number
              won:
                type: number
            required: [date, wagered, won]
      required: [totalBets, totalWagered, totalWon, winRate, profitLoss]
    ReferralInfo:
      type: object
      properties:
        referralCode:
          type: string
        totalReferrals:
          type: integer
        totalEarnings:
          type: number
        referrals:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              date:
                type: string
              earning:
                type: number
            required: [name, date, earning]
      required: [referralCode, totalReferrals, totalEarnings, referrals]
    Market:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        openTime:
          type: string
        closeTime:
          type: string
        status:
          type: string
        minBet:
          type: number
        maxBet:
          type: number
        payoutRatio:
          type: number
        isBettingOpen:
          type: boolean
        latestResult:
          $ref: "#/components/schemas/Result"
      required: [id, name, openTime, closeTime, status, minBet, maxBet, isBettingOpen]
...
    Bet:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        marketId:
          type: string
        marketName:
          type: string
        gameType:
          type: string
          enum: [jantri, open, crossing, no_to_no]
        numbers:
          type: array
          items:
            type: object
            properties:
              number:
                type: string
              amount:
                type: number
            required: [number, amount]
        totalAmount:
          type: number
        status:
          type: string
          enum: [pending, won, lost, cancelled]
        winAmount:
          type: number
        createdAt:
          type: string
        settledAt:
          type: string
      required: [id, marketId, gameType, numbers, totalAmount, status, createdAt]
...
    Result:
      type: object
      properties:
        id:
          type: string
        marketId:
          type: string
        marketName:
          type: string
        date:
          type: string
        openNumber:
          type: string
        closeNumber:
          type: string
        jodiNumber:
          type: string
        status:
          type: string
          enum: [scheduled, live, completed]
        createdAt:
          type: string
      required: [id, marketId, date, status]
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, bet, win, referral, commission, adjustment]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, rejected, cancelled]
        method:
          type: string
        utrNumber:
          type: string
        screenshotUrl:
          type: string
        adminNote:
          type: string
        bankDetails:
          type: object
          properties:
            accountNumber:
              type: string
            ifscCode:
              type: string
            accountHolderName:
              type: string
            upiId:
              type: string
        createdAt:
          type: string
        processedAt:
          type: string
      required: [id, type, amount, status, createdAt]
    AddMoneyRequest:
      type: object
      properties:
        amount:
          type: number
        utrNumber:
          type: string
        screenshotUrl:
          type: string
      required: [amount, utrNumber]
    WithdrawalRequest:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [bank, upi]
        bankAccountNumber:
          type: string
        ifscCode:
          type: string
        accountHolderName:
          type: string
        upiId:
          type: string
      required: [amount, method]
    UpiDetails:
      type: object
      properties:
        upiId:
          type: string
        qrCodeUrl:
          type: string
      required: [upiId]
...
    Notification:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, title, message, type, isRead, createdAt]
...